// api/src/models.ts
export interface Bible {
    bible_id: number;
    abbreviation: string;
    title: string;
}

export interface Book {
    book_id: number;
    bible_id: number;
    abbreviation: string;
    book_name: string;
    order_index: number;
}

export interface Verse {
    verse_id: number;
    verse_numbers: Array<number>;
    verse_text: string;
    header_text?: string;
    book_id: number;
    chapter_number: number;
    notes?: Array<string>;
}

export interface CompareInput {
    bible_ids: Array<number>;
    book_abbreviation: string;
    chapter_number: number;
    verse_numbers: number[];
}

export interface CompareOutput {
    verse_id: number;
    verse_text: string;
    chapter_number: number;
    book_id: number;
    header_text: string;
    notes: Array<string>;
    verse_numbers: Array<number>;
    bible_id: number;
    book_name: string;
    order_index: number;
    title: string;
    bible_abbreviation: string;
    book_abbreviation: string;
}